requirements.txt
biopython>=1.79
pandas>=1.3.0
requests>=2.28.0
matplotlib>=3.5.0
numpy>=1.21.0
